Add support for rich `libdns` structs
This change implements support to return DNS API
responses not just as `libdns.RR`, but as rich
types with parsed contents, as
available/[recommended as of libdns
v1.0.0](https://github.com/libdns/googleclouddns/pull/9#issuecomment-2802965563).

It also adjusts an existing test case to check for
whether we do this, and fixes a bug in that test
where it would pass if the sought record was
actually missing.